+++ events/push.js
@@ -1,0 +1,6 @@
+const { broadcast } = require('./common');
+const pushService = require('../services/push');
+
+broadcast.on('participate', (room_number, message) => {
+  //push 실행부
+});
